requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.12.1
olefile==0.44
opencv-python==3.2.0.7
Pillow==4.1.1
PyYAML==3.12
six==1.10.0
torch==0.1.12.post2
torchvision==0.1.8
Werkzeug==0.12.2
